Feature/286-schedule-update-page-front

+++ backend/src/models/entities/feed-media.entity.ts
@@ -49,7 +49,7 @@ export class FeedMediaEntity extends DefaultEntity {
 	/**
 	 * 미디어파일의 정렬 순서로, 오름차순 정렬된다.
 	 */
-	@Column('numeric', { name: 'position', default: 0 })
+	@Column('int4', { name: 'position', default: 0 })
 	@ApiProperty()
 	@IsNotEmpty({
 		message: notEmptyValidationMessage,
